requirements.txt
werkzeug
numpy
gunicorn
Flask
h5py
tensorflow-cpu
pillow
